Intervals with double as number type
